feature: Move Naja to `devDependencies`
- Naja has been moved from `dependencies` to `devDependencies`. The library doesn't import Naja's code directly anymore, but only its types.
- `ControlManager` is no longer exported as an instance, only as a class. This is a BC break, as you have to change your project code and instantiate it yourself.
- The `Control` interface is now exported from a different path. The path `@peckadesign/pd-naja/utils` has been changed to `@peckadesign/pd-naja/types`.
