+++ package.json
@@ -28,9 +28,6 @@
 		"url": "https://github.com/peckadesign/pd-naja/issues"
 	},
 	"homepage": "https://github.com/peckadesign/pd-naja#readme",
-	"dependencies": {
-		"naja": "^2.5.0"
-	},
 	"scripts": {
 		"build": "NODE_ENV=production rollup -c --bundleConfigAsCjs",
 		"lint": "NODE_ENV=development eslint src --ext .ts --max-warnings=0"
@@ -46,6 +43,7 @@
 		"eslint-config-prettier": "^8.8.0",
 		"eslint-config-typescript": "^3.0.0",
 		"eslint-plugin-prettier": "^4.2.1",
+		"naja": "2.5.0",
 		"prettier": "^2.8.8",
 		"rollup": "^3.21.6",
 		"typescript": "^5.0.4"
